Allow to set fields to `Default` when constructing struct from query
This commit adds a `#[diesel(default)]` field attribute to
`#[derive(Queryable)]`, that allows a user to set a field to a default
value instead of loading it from the database. This is usefull if the
same struct is used to serve some API that requires additional
information to be added.