Add a trait to abstract away the access to foreign key fields in belongs_to
This allows to use wrapper constructs as foreign key fields. Without
this change a foreign key is required to have "the same" type as the
primary key of the parent struct. After this change it is possible to
do some kind of transformation to get a compatible type out of the
wrapper struct