Try to cleanup mysql type metadata implementation
This commit adds a fake type layer above the types exposed by the
mysql wire protocol as some of them are unused, can be used in the same
context and some additional information are required to be read from
the provided flags integer. I do not want to expose such a mess to
potential users, better expose an clear type enum with all currently
supported variants in diesel. If we hit later something that isn't
supported yet we can just add it with an additional entry, because
this enum is `#[non_exhaustive]`