Improve the sqlite bind collector to skip even more copies
This improves the sqlite bind collector to skip even more data copies,
by setting the sqlite bind mode to `STATIC`. This prevents that sqlite
creates a internal copy of the bind data, which means we must ensure
that the data live at least as long as the buffer is bound to the
statement. This requires some unsafe hackery to ensure that all data
remain avaible. We ensure this by binding all information to
`StatementUse` itself, which lives as long as the user can access the
executed query. If `StatementUse` is dropped we cleanup the
corresponding buffers.